Extract StakeRewardCalculator to check rewards outside of action
